[KBSWITCH][INTERNAT][BOOTDATA][MEDIA] Rename 'kbswitch' and 'kbsdll' as 'internat' and 'indicdll'
