+++ base/applications/CMakeLists.txt
@@ -17,7 +17,7 @@ add_subdirectory(fontview)
 add_subdirectory(games)
 add_subdirectory(hh)
 add_subdirectory(iexplore)
-add_subdirectory(kbswitch)
+add_subdirectory(internat)
 add_subdirectory(logoff)
 add_subdirectory(magnify)
 add_subdirectory(mmc)
