+++ base/applications/internat/indicdll/indicdll.rc
@@ -1,5 +1,5 @@
 #define REACTOS_VERSION_DLL
 #define REACTOS_STR_FILE_DESCRIPTION  "ReactOS Keyboard Layout Switcher"
-#define REACTOS_STR_INTERNAL_NAME     "kbsdll"
-#define REACTOS_STR_ORIGINAL_FILENAME "kbsdll.dll"
+#define REACTOS_STR_INTERNAL_NAME     "indicdll"
+#define REACTOS_STR_ORIGINAL_FILENAME "indicdll.dll"
 #include <reactos/version.rc>
