+++ base/applications/internat/indicdll/indicdll.spec
