+++ base/applications/internat/internat.rc
@@ -5,11 +5,11 @@
 LANGUAGE LANG_NEUTRAL, SUBLANG_NEUTRAL
 
 #define REACTOS_STR_FILE_DESCRIPTION  "ReactOS Keyboard Layout Switcher"
-#define REACTOS_STR_INTERNAL_NAME     "kbswitch"
-#define REACTOS_STR_ORIGINAL_FILENAME "kbswitch.exe"
+#define REACTOS_STR_INTERNAL_NAME     "internat"
+#define REACTOS_STR_ORIGINAL_FILENAME "internat.exe"
 #include <reactos/version.rc>
 
-IDI_MAIN ICON "res/kbswitch.ico"
+IDI_MAIN ICON "res/internat.ico"
 
 /* UTF-8 */
 #pragma code_page(65001)
