+++ base/applications/internat/lang/bg-BG.rc
