+++ base/applications/internat/lang/cs-CZ.rc
@@ -1,4 +1,4 @@
-/* FILE:       applications/kbswitch/lang/cs-CZ.rc
+/* FILE:       base/applications/internat/lang/cs-CZ.rc
  * TRANSLATOR: Radek Liska aka Black_Fox (radekliska at gmail dot com)
  * UPDATED:    2008-04-25
  */
