+++ base/applications/internat/lang/de-DE.rc
