+++ base/applications/internat/lang/en-US.rc
