+++ base/applications/internat/lang/es-ES.rc
