+++ base/applications/internat/lang/et-EE.rc
