+++ base/applications/internat/lang/fr-FR.rc
