+++ base/applications/internat/lang/he-IL.rc
