+++ base/applications/internat/lang/it-IT.rc
