+++ base/applications/internat/lang/ja-JP.rc
