+++ base/applications/internat/lang/lt-LT.rc
