+++ base/applications/internat/lang/no-NO.rc
