+++ base/applications/internat/lang/pl-PL.rc
