+++ base/applications/internat/lang/pt-BR.rc
