+++ base/applications/internat/lang/pt-PT.rc
