+++ base/applications/internat/lang/ro-RO.rc
@@ -1,5 +1,5 @@
 /*
- * FILE:       base/applications/kbswitch/lang/ro-RO.rc
+ * FILE:       base/applications/internat/lang/ro-RO.rc
  *             ReactOS Project (https://reactos.org)
  * TRANSLATOR: Fulea Ștefan (PM on ReactOS Forum at fulea.stefan)
  * CHANGE LOG: 2011-08-21  initial translation
