+++ base/applications/internat/lang/ru-RU.rc
