+++ base/applications/internat/lang/sk-SK.rc
