+++ base/applications/internat/lang/sq-AL.rc
