+++ base/applications/internat/lang/sv-SE.rc
