+++ base/applications/internat/lang/tr-TR.rc
