+++ base/applications/internat/lang/uk-UA.rc
@@ -1,7 +1,7 @@
 /*
  * PROJECT:    Keyboard Layout Switcher
  * LICENSE:    GPL - See COPYING in the top level directory
- * FILE:       base/applications/kbswitch/lang/uk-UA.rc
+ * FILE:       base/applications/internat/lang/uk-UA.rc
  * PURPOSE:    Ukraianian Language File for kbswitch
  * TRANSLATOR: Artem Reznikov
  */
