+++ base/applications/internat/lang/zh-HK.rc
