+++ base/applications/internat/lang/zh-TW.rc
@@ -1,7 +1,7 @@
 /*
  * PROJECT:    ReactOS KBSwitch
  * LICENSE:    GPL - See COPYING in the top level directory
- * FILE:       base/applications/kbswitch/lang/zh-TW.rc
+ * FILE:       base/applications/internat/lang/zh-TW.rc
  * PURPOSE:    Chinese (Traditional) Language File for KBSwitch
  * TRANSLATOR: Elton Chung aka MfldElton <elton328@gmail.com>
  */
