+++ base/applications/internat/resource.h
