+++ boot/bootdata/hivedef.inf
@@ -1928,7 +1928,7 @@ HKCU,"SOFTWARE\Microsoft\Windows\CurrentVersion\Telephony",,0x00000012
 HKCU,"SOFTWARE\Microsoft\Windows\CurrentVersion\Uninstall",,0x00000012
 
 ; Keyboard layout switcher
-HKCU,"SOFTWARE\Microsoft\Windows\CurrentVersion\Run","kbswitch.exe",0x00020002,"%SystemRoot%\system32\kbswitch.exe"
+HKCU,"SOFTWARE\Microsoft\Windows\CurrentVersion\Run","internat.exe",0x00020002,"%SystemRoot%\system32\internat.exe"
 
 HKCU,"SOFTWARE\Microsoft\Windows NT",,0x00000012
 HKCU,"SOFTWARE\Microsoft\Windows NT\CurrentVersion",,0x00000012
